Fall back to Event.target if port is undefined
I am considering removing the custom SerialConnectionEvent interface and
firing 'connect' and 'disconnect' events directly at the SerialPort
instance that has changed state. This means that the SerialPort will be
available as the target attribute rather than the custom port attribute.

This changes makes the app compatible with either implementation.