Use a BYOB reader if supported (#40)
* Use a BYOB reader if supported

This change adds code to detect when the ReadableStream is a byte stream
and so a BYOB reader can be created. In this case the same ArrayBuffer
memory is reused for every call to read().

This change required updating TypeScript and explicitly including the
@types/web package.

* Fix BYOB view creation

Also improves resiliancy to error loops by waiting for the terminal to
render the error message before trying again. Otherwise a tight loop can
be created which blocks the entire UI.